+++ data/friendships.json
@@ -13,5 +13,10 @@
         "userId1": "U1",
         "userId2": "U2",
         "status": "pending"
+    },
+    {
+        "userId1": "U1",
+        "userId2": "U3",
+        "status": "pending"
     }
 ]
