+++ src/main/java/backend/friendship/IFriendShipValidation.java
@@ -1,7 +1,6 @@
-package backend;
+package backend.friendship;
 
 import org.json.JSONArray;
-import org.json.JSONObject;
 
 public interface IFriendShipValidation {
     public boolean checkDuplicates(String userId1, String userId2, JSONArray friendships);
