+++ src/main/java/backend/friendship/ILoadFriendShips.java
@@ -1,4 +1,4 @@
-package backend;
+package backend.friendship;
 
 import org.json.JSONArray;
