+++ src/main/java/backend/profile/ILoadProfiles.java
@@ -1,4 +1,4 @@
-package backend;
+package backend.profile;
 
 import org.json.JSONArray;
