+++ src/main/java/backend/profile/UpdateProfile.java
@@ -1,4 +1,4 @@
-package backend;
+package backend.profile;
 
 import org.json.JSONArray;
 import org.json.JSONObject;
