+++ src/main/java/backend/user/IAddUser.java
@@ -1,4 +1,4 @@
-package backend;
+package backend.user;
 
 public interface IAddUser {
     public void addUser(String username, String password , String email , String dob);
