+++ src/main/java/backend/user/PasswordUtils.java
@@ -1,9 +1,9 @@
-package backend;
+package backend.user;
 
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
 import java.util.Base64;
-import java.io.FileWriter;
+
 import org.json.JSONArray;
 import org.json.JSONObject;
