+++ src/main/java/backend/user/UpdateUser.java
@@ -1,4 +1,4 @@
-package backend;
+package backend.user;
 
 import org.json.JSONArray;
 import org.json.JSONObject;
