+++ src/main/java/frontend/friendshipManagement/FriendController.java
@@ -1,14 +1,13 @@
-package frontend;
+package frontend.friendshipManagement;
 
-import backend.*;
+import backend.friendship.FriendShip;
+import backend.friendship.FriendShipFactory;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
 import javafx.scene.control.Alert;
 import javafx.scene.control.ListView;
 import javafx.scene.control.TextField;
-import org.json.JSONArray;
-import org.json.JSONObject;
 
 import java.util.List;
